Refactor package structure for EvmFeeModule classes
remove old EvmTransactionFeeService
