+++ app/src/main/java/io/horizontalsystems/bankwallet/core/Warning.kt
@@ -1,0 +1,3 @@
+package io.horizontalsystems.bankwallet.core
+
+abstract class Warning
