+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/evmfee/eip1559/Eip1559FeeSettingsViewModel.kt
@@ -1,0 +1,5 @@
+package io.horizontalsystems.bankwallet.modules.evmfee.eip1559
+
+class Eip1559FeeSettingsViewModel {
+
+}
