+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/evmfee/eip1559/Eip1559GasPriceService.kt
@@ -1,0 +1,4 @@
+package io.horizontalsystems.bankwallet.modules.evmfee.eip1559
+
+class Eip1559GasPriceService {
+}
