--- app/src/main/java/io/horizontalsystems/bankwallet/modules/sendevmtransaction/feesettings/Eip1559GasPriceService.kt
@@ -1,4 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.sendevmtransaction.feesettings
-
-class Eip1559GasPriceService {
-}
