--- app/src/main/java/io/horizontalsystems/bankwallet/modules/sendevmtransaction/feesettings/Eip1559GasPriceViewModel.kt
@@ -1,5 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.sendevmtransaction.feesettings
-
-class Eip1559GasPriceViewModel {
-
-}
