+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/oneinch/OneInchSwapViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModel
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.EvmError
 import io.horizontalsystems.bankwallet.core.convertedError
-import io.horizontalsystems.bankwallet.core.ethereum.EvmTransactionFeeService
 import io.horizontalsystems.bankwallet.core.providers.Translator
+import io.horizontalsystems.bankwallet.modules.evmfee.GasDataError
 import io.horizontalsystems.bankwallet.modules.swap.SwapMainModule
 import io.horizontalsystems.bankwallet.modules.swap.SwapMainModule.ApproveStep
 import io.horizontalsystems.bankwallet.modules.swap.SwapMainModule.SwapError
@@ -139,7 +139,7 @@ class OneInchSwapViewModel(
 
     private fun sync(errors: List<Throwable>) {
         val filtered =
-            errors.filter { it !is EvmTransactionFeeService.GasDataError && it !is SwapError }
+            errors.filter { it !is GasDataError && it !is SwapError }
         swapErrorLiveData.postValue(filtered.firstOrNull()?.let { convert(it) })
 
         syncState()
